feat(core): Give access to the `ActiveEventLoop` in the window attributes hook
